Create GUISaveAs
1. createNewWindow function serves as the function that replaces the textArea when a new file is being named. I do believe, in this function, a "text" string stores the value of textArea in order to place it in the file (see "save" discussion)

createDirectory should make a directory/folder in your personal documents photo called "Entries", in which you can find your text files. These are stored in an arrayList but are not pulled from the arrayList, rather simply by opening the file

2. Detecting button press of either save or cancel prompts this menu.

If the save button is pressed:
1. The header is collected from the TextField in the New File panel.
2. It is used to create a new file in the "Entries" directory. This file now serves as the current "openFile".
3. Using FileWriter, the "text" (storing textArea) writes to this new file, and uses the same date saving system as previously discussed.
4. Text fields are reset to default.
5. Panel switches back to ScrollPane.

Last change: I made the textArea clear when "Open" was pressed. I figured it was best to present the file itself instead of some old entry being overlapped by a new one, especially for purposes of overwriting a file